Move reworked flow tests to the test/flows directory
As these tests are caught in-between integration and unit tests it
doesn't seem ideal to put them into a test/integration or a test/unit
directory as you may expect them to be present in the other.

As part of this I've renamed them all to have a suffix of
ExperimentalFlow to reflect that they are all experimental flow tests.

This also moves the ExperimentalFlowTestHelper to the support directory
which seemed a more idiomatic location than having the helper file
amongst the tests.

The future steps planned for this directory move would be:

- Move the flows themselves from lib/smart_answer_flows to app/flows
- Remove the Experimental word from the name.
