+++ test/flows/towing_rules_experimental_flow_test.rb
@@ -1,9 +1,9 @@
-require_relative "../../test_helper"
-require_relative "experimental_flow_test_helper"
+require "test_helper"
+require "support/experimental_flow_test_helper"
 
 require "smart_answer_flows/towing-rules"
 
-class TowingRulesReworkedTest < ActiveSupport::TestCase
+class TowingRulesExperimentalFlowTest < ActiveSupport::TestCase
   include ExperimentalFlowTestHelper
 
   setup { testing_flow SmartAnswer::TowingRulesFlow }
